+++ src/main/java/com/mycompany/model/PessoaJuridica.java
@@ -27,7 +27,7 @@ public class PessoaJuridica  implements Serializable {
     @Column(nullable=false, length=200)
     private String Telefone;
     @Column(nullable=false, length=200)
-    private Long rendaMensal;
+    private Double rendaMensal;
     @Column(nullable=false, length=200)
     private String cidade;
     @Column(nullable=false, length=200)
@@ -138,11 +138,11 @@ public class PessoaJuridica  implements Serializable {
         this.conta = conta;
     }
 
-    public Long getRendaMensal() {
+    public Double getRendaMensal() {
         return rendaMensal;
     }
 
-    public void setRendaMensal(Long rendaMensal) {
+    public void setRendaMensal(Double rendaMensal) {
         this.rendaMensal = rendaMensal;
     }
